style: Move Custom Bottom Navigation Bar Shape From widgets To shapes
